# runtime: Use ainur.StreamReader instead of io.RuneReader