- Use ainur.StreamReader instead of io.RuneReader
- Fix tests

### Benchmark


<details><summary>Old</summary>
<p>

```text
oos: linux
goarch: amd64
pkg: github.com/parca-dev/parca-agent/pkg/runtime/nodejs
cpu: AMD Ryzen 9 5950X 16-Core Processor            
Benchmark_scanVersionBytes-32                100          11508930 ns/op           32162 B/op        178 allocs/op
Benchmark_scanVersionBytes-32                102          11422398 ns/op           32672 B/op        178 allocs/op
Benchmark_scanVersionBytes-32                100          11774011 ns/op           32991 B/op        178 allocs/op
Benchmark_scanVersionBytes-32                100          11629588 ns/op           32162 B/op        178 allocs/op
Benchmark_scanVersionBytes-32                102          12016436 ns/op           33038 B/op        178 allocs/op
Benchmark_scanVersionBytes-32                103          11492685 ns/op           32602 B/op        178 allocs/op
Benchmark_scanVersionBytes-32                102          11671819 ns/op           33038 B/op        178 allocs/op
Benchmark_scanVersionBytes-32                100          11675831 ns/op           32601 B/op        178 allocs/op
Benchmark_scanVersionBytes-32                100          11757764 ns/op           32601 B/op        178 allocs/op
Benchmark_scanVersionBytes-32                103          11512288 ns/op           32602 B/op        178 allocs/op
PASS
ok      github.com/parca-dev/parca-agent/pkg/runtime/nodejs     15.678s
```

</p>
</details> 

<details><summary>New</summary>
<p>

```text
goos: linux
goarch: amd64
pkg: github.com/parca-dev/parca-agent/pkg/runtime/nodejs
cpu: AMD Ryzen 9 5950X 16-Core Processor            
Benchmark_scanVersionBytes-32               1958            656223 ns/op           32144 B/op        174 allocs/op
Benchmark_scanVersionBytes-32               1585            727980 ns/op           32031 B/op        174 allocs/op
Benchmark_scanVersionBytes-32               2030            689447 ns/op           32172 B/op        174 allocs/op
Benchmark_scanVersionBytes-32               1796            670143 ns/op           32121 B/op        174 allocs/op
Benchmark_scanVersionBytes-32               1525            715696 ns/op           32209 B/op        174 allocs/op
Benchmark_scanVersionBytes-32               1504            715440 ns/op           32136 B/op        174 allocs/op
Benchmark_scanVersionBytes-32               1597            676401 ns/op           32100 B/op        174 allocs/op
Benchmark_scanVersionBytes-32               1540            695197 ns/op           32181 B/op        174 allocs/op
Benchmark_scanVersionBytes-32               1615            718022 ns/op           32097 B/op        174 allocs/op
Benchmark_scanVersionBytes-32               1591            688324 ns/op           32242 B/op        174 allocs/op
PASS
ok      github.com/parca-dev/parca-agent/pkg/runtime/nodejs     13.306s
```

</p>
</details> 

**Comparison**
```text
goos: linux
goarch: amd64
pkg: github.com/parca-dev/parca-agent/pkg/runtime/nodejs
cpu: AMD Ryzen 9 5950X 16-Core Processor
                     │    old.txt    │               new.txt               │
                     │    sec/op     │   sec/op     vs base                │
_scanVersionBytes-32   11650.7µ ± 1%   692.3µ ± 4%  -94.06% (p=0.000 n=10)

                     │   old.txt    │               new.txt               │
                     │     B/op     │     B/op      vs base               │
_scanVersionBytes-32   31.84Ki ± 1%   31.39Ki ± 0%  -1.42% (p=0.001 n=10)

                     │  old.txt   │              new.txt              │
                     │ allocs/op  │ allocs/op   vs base               │
_scanVersionBytes-32   178.0 ± 0%   174.0 ± 0%  -2.25% (p=0.000 n=10)
```

### TODO

- [ ] Remove checked in binary

Signed-off-by: Kemal Akkoyun <kakkoyun@gmail.com>